src/validation/validators/email/emailValidation.ts
import { InvalidFieldError } from '@/validation/errors'
import { FieldValidation } from '@/validation/protocols'

export class EmailValidation implements FieldValidation {
  #regex = /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/

  constructor (readonly name: string) {}

  validate (input: object) {
    const value = input[this.name]

    return !value || this.#regex.test(value) ? null : new InvalidFieldError(this.name)
  }
}
